refactor(routes): break out device-related routes to a separate module
