+++ test/local/routes/account.js
@@ -34,17 +34,10 @@ var makeRoutes = function (options, requireMocks) {
     idle: 500,
     lifetime: 30
   }
-  config.i18n = {
-    supportedLanguages: ['en'],
-    defaultLanguage: 'en'
-  }
   config.lastAccessTimeUpdates = {}
   config.signinConfirmation = config.signinConfirmation || {}
   config.signinUnblock = config.signinUnblock || {}
   config.secondaryEmail = config.secondaryEmail || {}
-  config.push = {
-    allowedServerRegex: /^https:\/\/updates\.push\.services\.mozilla\.com(\/.*)?$/
-  }
 
   var log = options.log || mocks.mockLog()
   var Password = options.Password || require('../../../lib/crypto/password')(log, config)
@@ -63,8 +56,7 @@ var makeRoutes = function (options, requireMocks) {
     config,
     customs,
     checkPassword,
-    push,
-    options.devices || require('../../../lib/devices')(log, db, push)
+    push
   )
 }
 
@@ -1199,69 +1191,3 @@ describe('/account/destroy', function () {
   })
 })
 
-
-describe('/account/sessions', function () {
-  var testSession = {
-    tokenId: 'foo',
-    uid: '010',
-    createdAt: Date.now(),
-    uaOS: 'Windows',
-    uaOSVersion: '10',
-    uaDeviceType: 'desktop',
-    lastAccessTime: Date.now(),
-
-    uaBrowser: 'Firefox',
-    uaBrowserVersion: '50',
-
-    deviceId: 'deviceId',
-    deviceName: 'deviceName',
-    deviceType: 'desktop',
-    deviceCreatedAt: Date.now(),
-    devicePushCallback: 'foo',
-    devicePushPublicKey: 'bar',
-    devicePushAuthKey: 'authKey',
-  }
-  var mockDB = mocks.mockDB()
-  mockDB.sessions = sinon.spy(function () {
-    return P.resolve([
-      Object.assign({}, testSession)
-    ])
-  })
-
-  var mockRequest = mocks.mockRequest({
-    credentials: {
-      uid: hexString(16),
-      tokenId: hexString(16)
-    },
-    payload: {}
-  })
-
-  var accountRoutes = makeRoutes({
-    db: mockDB
-  })
-
-
-  it('should list account sessions', () => {
-    var route = getRoute(accountRoutes, '/account/sessions')
-
-    return runTest(route, mockRequest, function (res) {
-      assert.equal(res.length, 1)
-      assert.equal(Object.keys(res[0]).length, 13)
-      var s = res[0]
-      assert.equal(s.id, testSession.tokenId)
-      assert.equal(s.deviceName, testSession.deviceName)
-      assert.equal(s.deviceType, testSession.deviceType)
-      assert.equal(s.devicePushCallback, testSession.devicePushCallback)
-      assert.equal(s.devicePushPublicKey, testSession.devicePushPublicKey)
-      assert.equal(s.devicePushAuthKey, testSession.devicePushAuthKey)
-      assert.equal(s.id, testSession.tokenId)
-      assert.equal(s.isCurrentDevice, false)
-      assert.equal(s.isDevice, true)
-      assert.equal(s.lastAccessTimeFormatted, 'a few seconds ago')
-      assert.equal(s.userAgent, 'Firefox 50')
-      assert.equal(s.os, 'Windows')
-    })
-
-  })
-
-})
